+++ coherence/src/coherence/bus/MessageMetadata.scala
@@ -2,4 +2,4 @@ package coherence.bus
 
 import coherence.Unit.Word
 
-case class MessageMetadata[Message](message: Message, address: Int, size: Word)
+case class MessageMetadata[Message](message: Message, address: Long, size: Word)
