+++ coherence/src/coherence/devices/Memory.scala
@@ -3,11 +3,11 @@ package coherence.devices
 import scala.collection.mutable
 
 class Memory extends Device {
-  private[this] var currentCycle: Int = 0
+  private[this] var currentCycle: Long = 0
 
   // A map from (op, address) to finishedCycle
   private[this] val requests =
-    mutable.Map[(MemoryDelegate, MemoryOp, Int), Int]()
+    mutable.Map[(MemoryDelegate, MemoryOp, Long), Long]()
 
   def cycle(): Unit = {
     currentCycle += 1
@@ -20,11 +20,11 @@ class Memory extends Device {
       }
   }
 
-  def operate(sender: MemoryDelegate, op: MemoryOp, address: Int): Unit = {
+  def operate(sender: MemoryDelegate, op: MemoryOp, address: Long): Unit = {
     requests.put((sender, op, address), currentCycle + Memory.Latency)
   }
 
-  def cancel(sender: MemoryDelegate, op: MemoryOp, address: Int): Unit = {
+  def cancel(sender: MemoryDelegate, op: MemoryOp, address: Long): Unit = {
     requests.remove((sender, op, address))
   }
 }
