+++ coherence/src/coherence/devices/MemoryDelegate.scala
@@ -1,5 +1,5 @@
 package coherence.devices
 
 trait MemoryDelegate {
-  def memoryOpCompleted(op: MemoryOp, address: Int)
+  def memoryOpCompleted(op: MemoryOp, address: Long)
 }
