+++ coherence/src/coherence/mesi/Message.scala
@@ -2,6 +2,6 @@ package coherence.mesi
 
 sealed trait Message
 
-case class BusRd(address: Int) extends Message
-case class BusRdX(address: Int) extends Message
-case class FlushOpt(address: Int) extends Message
+case class BusRd(address: Long) extends Message
+case class BusRdX(address: Long) extends Message
+case class FlushOpt(address: Long) extends Message
